requirements.txt
aws-lambda-powertools==3.1.0
boto3==1.35.44
boto3-stubs==1.35.44
botocore==1.35.44
botocore-stubs==1.35.44
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
freezegun==1.5.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
tomli==2.0.2
types-awscrt==0.22.0
types-boto3==1.0.2
types-s3transfer==0.10.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
